Fix to properly wipe filesystems larger than 4 gigabytes.
Pass the wipe size to the wipe function in a 64 bit int.

+++ ext4_utils/wipe.h
@@ -17,6 +17,6 @@
 #ifndef _WIPE_H_
 #define _WIPE_H_
 
-int wipe_block_device(int fd, int len);
+int wipe_block_device(int fd, s64 len);
 
 #endif
